Update SievePrimesAlgorithm.java
prime[int] == true (lines 15 & 37(now 38 with try block)) can just read prime[n] in the if statement because it is itself a boolean.

Also if someone were to enter an int greater than N they'll get an array out of bounds exception. Here it is caught and printed, but you could just return false as with anything less than or equal to 1.